Add .qmltypes files
.qmltypes files are used by e.g. Qt Creator to provide code completion etc
for QML types defined in plugins. The added files where generated with
qmlplugindump, e.g.

  qmlplugindump QtQuick.Particles 2.0 > plugins.qmltypes

(ideally make qmltypes would do the same, but it's not working as expected
right now).

The exception is the QtQuick plugin: Here the file is generated with

  qmlplugindump -builtins > plugins.qmltypes

and contains also the types registered in the qml library itself.

Change-Id: I1a6b6641cb5ec3ecffe08e8926d8c1bc082ae6de
Reviewed-by: Christian Kamm <christian.d.kamm@nokia.com>