Add "showFiles" property to FolderListModel
This property, if set to "false", causes the FolderListModel not to show
regular files. This can be useful when someone wants to list only the
directories; for instance, one can have a ListView showing only the
directories (with "showDirs: true" and "showFiles: false") and another
for showing the directory contents ("showFiles: true" and "showDirs:
false").

Since by default this property is "true", the behavior of applications
who don't use this property is unchanged.

Change-Id: I6eaa9c4a6a582ea81fa07bad129a4c9b15fdcaea
Reviewed-by: Alan Alpert <aalpert@blackberry.com>